tests: use fail point
Remove the hacky function pointer, and use failpoint to do the precise
control.
